+++ spec/kettle/dev/release_cli_spec.rb
@@ -1087,7 +1087,7 @@ RSpec.describe Kettle::Dev::ReleaseCLI do
           LOCK
 
           expect(local_cli).to receive(:run_cmd!).with(
-            a_string_matching(/KETTLE_DEV_DEV=false.*BUNDLE_GEMFILE=.*Gemfile.*bundle lock --update --add-checksums/)
+            a_string_matching(/KETTLE_DEV_DEV=false.*BUNDLE_GEMFILE=.*Gemfile.*bundle lock .*--update --add-checksums/)
           ) do
             File.write(File.join(root, "Gemfile.lock"), <<~LOCK)
               GEM
@@ -1145,7 +1145,7 @@ RSpec.describe Kettle::Dev::ReleaseCLI do
 
           expect(local_cli).to receive(:run_cmd!).with(
             a_string_matching(
-              /BUNDLE_GEMFILE=.*Appraisal\.root\.gemfile.*BUNDLE_LOCKFILE=.*Appraisal\.root\.gemfile\.lock.*bundle lock --update --add-checksums/
+              /BUNDLE_GEMFILE=.*Appraisal\.root\.gemfile.*BUNDLE_LOCKFILE=.*Appraisal\.root\.gemfile\.lock.*bundle lock .*--update --add-checksums/
             )
           ).ordered do
             File.write(File.join(root, "Appraisal.root.gemfile.lock"), <<~LOCK)
@@ -1162,7 +1162,7 @@ RSpec.describe Kettle::Dev::ReleaseCLI do
             LOCK
           end
           expect(local_cli).to receive(:run_cmd!).with(
-            a_string_matching(/BUNDLE_GEMFILE=.*Gemfile.*BUNDLE_LOCKFILE=.*Gemfile\.lock.*bundle lock --update --add-checksums/)
+            a_string_matching(/BUNDLE_GEMFILE=.*Gemfile.*BUNDLE_LOCKFILE=.*Gemfile\.lock.*bundle lock .*--update --add-checksums/)
           ).ordered do
             File.write(File.join(root, "Gemfile.lock"), <<~LOCK)
               GEM
@@ -1304,7 +1304,7 @@ RSpec.describe Kettle::Dev::ReleaseCLI do
           LOCK
 
           expect(local_cli).to receive(:run_cmd!).with(
-            a_string_matching(/bundle lock --update --add-checksums/)
+            a_string_matching(/bundle lock .*--update --add-checksums/)
           ) do
             File.write(File.join(root, "Gemfile.lock"), <<~LOCK)
               GEM
@@ -1358,7 +1358,7 @@ RSpec.describe Kettle::Dev::ReleaseCLI do
           LOCK
 
           expect(local_cli).to receive(:run_cmd!).with(
-            a_string_matching(/bundle lock --update --add-checksums/)
+            a_string_matching(/bundle lock .*--update --add-checksums/)
           ) do
             File.write(File.join(root, "Gemfile.lock"), <<~LOCK)
               GEM